Refactor web application to new structure following best practices and enhancing UX
