+++ settle_up/__init__.py
@@ -1,0 +1,5 @@
+"""
+Settle Up Analyzer - A web application for analyzing Settle Up expenses.
+"""
+
+__version__ = "0.1.0"
