+++ settle_up/core/__init__.py
@@ -1,0 +1,1 @@
+"""Core functionality package for Settle Up Analyzer."""
