+++ settle_up/core/analyzer.py
@@ -1,36 +1,31 @@
+"""Data analysis functionality for Settle Up expenses."""
 import os
 from typing import Optional
 from functools import lru_cache
-from dataclasses import dataclass
 
 import pandas as pd
 from pandas import DataFrame
-from datetime import datetime
-from config import logger, settings
 
-
-def get_plot_filename(workdir: str, filename: str) -> str:
-    return os.path.join(workdir, filename)
-
-
-@dataclass
-class AnalyzerData:
-    """Container for commonly used dataframes to avoid recomputation"""
-    df: DataFrame
-    df_expenses: DataFrame
-    df_ingress: DataFrame
-    df_summary: DataFrame
+from .models import AnalyzerData
 
 
 class Analyzer:
-    def __init__(self) -> None:
-        self.workdir: str = settings["workdir"]
-        self.plotdir: str = os.path.join(self.workdir, "plots")
+    """Analyzes processed Settle Up expense data."""
+
+    def __init__(self, workdir: str, excel_filename: str) -> None:
+        """Initialize the analyzer.
+        
+        Args:
+            workdir: Directory containing the Excel file
+            excel_filename: Name of the Excel file to analyze
+        """
+        self.workdir = workdir
+        self.excel_filename = excel_filename
         self._data: Optional[AnalyzerData] = None
 
     @property
     def data(self) -> AnalyzerData:
-        """Cached access to commonly used dataframes"""
+        """Cached access to commonly used dataframes."""
         if self._data is None:
             df = self._read_excel()
             df_expenses = df.loc[df["Type"] == "Expense"]
@@ -40,10 +35,8 @@ class Analyzer:
         return self._data
 
     def _read_excel(self) -> DataFrame:
-        """Read and preprocess the Excel file"""
-        excel_name = settings["expenses_excel_filename"]
-        filepath = os.path.join(self.workdir, excel_name)
-        logger.debug(f"Reading excel {filepath}")
+        """Read and preprocess the Excel file."""
+        filepath = os.path.join(self.workdir, self.excel_filename)
         
         df: DataFrame = pd.read_excel(
             filepath,
@@ -58,14 +51,14 @@ class Analyzer:
         return df
 
     def _calculate_summary(self, df_expenses: DataFrame, df_ingress: DataFrame) -> DataFrame:
-        """Calculate yearly summary statistics
+        """Calculate yearly summary statistics.
         
         Args:
             df_expenses: DataFrame containing expense records
             df_ingress: DataFrame containing ingress records
             
         Returns:
-            DataFrame with yearly summary statistics including expenses, ingress, and savings
+            DataFrame with yearly summary statistics
         """
         expenses_year = df_expenses.groupby("Year")["Amount"].sum()
         ingress_year = df_ingress.groupby("Year")["Amount"].sum()
@@ -81,24 +74,21 @@ class Analyzer:
 
     @lru_cache(maxsize=32)
     def month_summary(self, year: Optional[int] = None) -> DataFrame:
-        """Calculate monthly summary for a specific year or all years
+        """Calculate monthly summary for a specific year or all years.
         
         Args:
             year: Optional year to filter the data. If None, returns summary for all years.
             
         Returns:
-            DataFrame with monthly summary statistics including expenses, ingress, and savings
+            DataFrame with monthly summary statistics
         """
-        # Get cached dataframes
         df_ingress = self.data.df_ingress
         df_expenses = self.data.df_expenses
 
-        # Filter by year if specified
         if year:
             df_ingress = df_ingress.loc[df_ingress.Year == year]
             df_expenses = df_expenses.loc[df_expenses.Year == year]
 
-        # Calculate monthly summaries
         per_ingress = df_ingress.index.to_period("M")
         per_expenses = df_expenses.index.to_period("M")
         
@@ -115,3 +105,18 @@ class Analyzer:
         summary.index = summary.index.to_timestamp()
         
         return summary
+
+    def get_category_summary(self, year: Optional[int] = None) -> DataFrame:
+        """Get expense summary by category.
+        
+        Args:
+            year: Optional year to filter the data
+            
+        Returns:
+            DataFrame with category summary
+        """
+        df = self.data.df_expenses
+        if year:
+            df = df.loc[df.Year == year]
+            
+        return df.groupby("Category")["Amount"].agg(["sum", "count", "mean"]).round(2)
