+++ settle_up/web/__init__.py
@@ -1,0 +1,1 @@
+"""Web application package for Settle Up Analyzer."""
